Replaced regex used in reference replacement
Instead of creating a regex for each individual script created, now the
logic is that, for each script, just look for where a table without
reference is present, and use a map to replace it with the correct
reference. And if the table is not in the map, assume that is already
present somewhere else on the DBT project and just use the reference
macro either way, with that table name.

This speeds up SEVERELY the script.